requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
black==22.10.0
boto3==1.24.93
botocore==1.27.93
bravado==11.0.3
bravado-core==5.17.1
cachetools==5.2.0
cairocffi==1.4.0
CairoSVG==2.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chess==1.9.3
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.5
cryptography==38.0.1
cssselect2==0.7.0
cycler==0.11.0
datasets==2.6.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
docker==6.0.0
entrypoints==0.4
evaluate==0.3.0
exceptiongroup==1.0.1
executing==1.2.0
fabric==2.7.1
filelock==3.8.0
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.10.0
future==0.18.2
gin-config==0.5.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.13.0
google-cloud==0.34.0
huggingface-hub==0.10.1
idna==3.4
iniconfig==1.1.1
invoke==1.7.3
ipykernel==6.16.2
ipython==8.5.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonref==0.3.0
jsonschema==3.2.0
jupyter_client==7.4.4
jupyter_core==4.11.2
kiwisolver==1.4.4
kubernetes==24.2.0
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
meson==1.0.0
monotonic==1.6
mpmath==1.2.1
msgpack==1.0.4
multidict==6.0.2
multiprocess==0.70.13
munch==2.5.0
mypy-extensions==0.4.3
neptune-client==0.16.9
nest-asyncio==1.5.6
networkx==2.8.7
ninja==1.11.1
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.0
paramiko==2.11.0
parso==0.8.3
path==16.5.0
path.py==12.5.0
pathlib2==2.3.7.post1
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.5.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.18.1
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
pyzmq==24.0.1
QDarkStyle==3.1
QtAwesome==1.2.1
qtconsole==5.3.2
QtPy==2.2.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rfc3987==1.3.8
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.3
sente
simplejson==3.17.6
six==1.16.0
sklearn==0.0
smmap==5.0.0
stack-data==0.6.0
strict-rfc3339==0.7
swagger-spec-validator==2.7.6
termcolor==2.0.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.1
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
traitlets==5.5.0
transformers==4.23.1
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.1
xxhash==3.1.0
yarl==1.8.1
pydantic
backoff==2.2.1
rich==13.3.3